fix: Assign commits to versions following their commit graph
Before a commit was assigned the version/tag that happened next in the timeline. This leads to versions referencing commits that are not ancestors of it (and therefor not part of this version).

Instead the versions should contain only all commits that are ancestors, up until a new version is encountered.