feat: Support for additional template variables
Additional template variables for the Jinja template can be passed via CLI or can be defined in a configuration file.

To pass variables via CLI, the option "--template-var KEY VALUE" can be given multiple times.

To define variables in a configuration file, the variables can be defined in a "[template_vars]"  table.

[template_vars]
KEY = "VALUE"

All values are accessible in the template through the "template_vars" (dict) variable.