perf: Fix performance issue when assigning previous version
Assigning the previous version was checking commits multiple times, when ever a merge commit was encountered.
This leads to severe performance issues on a commit graph with lots of merge commits.

Instead of traversing the commit graph again, finding the previous version is now part of the commit to version grouping.